# app/api/unit.py
"""Units router (catalog) — backed by UnitService.

Catalog writes are admin/seed in principle; gating is deferred until auth
(SPEC.md "Authentication & authorization").
"""

from uuid import UUID

from fastapi import APIRouter, Depends, Query, Response, status
from sqlmodel import Session, SQLModel

from app.api.deps import get_current_admin, get_owned_by
from app.api.fields import Name, Stat, WriteSchema
from app.api.pagination import Page, PageParams, paginate
from app.core.db.connection import get_session
from app.core.services.service_unit import UnitService

router = APIRouter(prefix="/units", tags=["units"])


class Weapon_Read(SQLModel):
    id: UUID
    name: str
    category: str
    keywords: list[str]
    range_inches: int | None
    attacks: str
    weapon_skill: int
    strength: int
    armor_piercing: int
    damage: str


class Ability_Read(SQLModel):
    id: UUID
    name: str
    description: str


class Unit_Read(SQLModel):
    id: UUID
    unit_name: str
    faction_id: UUID
    subfaction_id: UUID | None
    movement: int
    toughness: int
    armor_save: int
    wounds: int
    invulnerable_save: int | None
    leadership: int
    objective_control: int
    points: int
    keywords: list[str]
    weapons: list[Weapon_Read] = []
    abilities: list[Ability_Read] = []


class Unit_Create(WriteSchema):
    faction_id: UUID
    unit_name: Name
    movement: Stat
    toughness: Stat
    armor_save: Stat
    wounds: Stat
    leadership: Stat
    objective_control: Stat
    points: Stat
    invulnerable_save: Stat | None = None
    subfaction_id: UUID | None = None
    keywords: list[str] | None = None


class Unit_Update(WriteSchema):
    unit_name: Name | None = None
    faction_id: UUID | None = None
    subfaction_id: UUID | None = None
    movement: Stat | None = None
    toughness: Stat | None = None
    armor_save: Stat | None = None
    wounds: Stat | None = None
    invulnerable_save: Stat | None = None
    leadership: Stat | None = None
    objective_control: Stat | None = None
    points: Stat | None = None
    keywords: list[str] | None = None


class UnitFacets(SQLModel):
    # Per-faction unit counts for the current filter, computed server-side in one
    # GROUP BY (see UnitService.faction_facets). `total` is the count across all
    # factions for the same filter. Replaces the client's download-and-count hack.
    total: int
    by_faction: dict[UUID, int]


class WeaponLink(WriteSchema):
    weapon_id: UUID


class AbilityLink(WriteSchema):
    ability_id: UUID


def get_unit_service(session: Session = Depends(get_session)) -> UnitService:
    return UnitService(session)


@router.post(
    "",
    response_model=Unit_Read,
    status_code=status.HTTP_201_CREATED,
    dependencies=[Depends(get_current_admin)],
)
def create_unit(payload: Unit_Create, service: UnitService = Depends(get_unit_service)) -> Unit_Read:
    return service.create_unit(**payload.model_dump())


@router.get("", response_model=Page[Unit_Read])
def list_units(
    faction_id: UUID | None = None,
    subfaction_id: UUID | None = None,
    q: str | None = Query(default=None, description="case-insensitive name search"),
    page: PageParams = Depends(),
    service: UnitService = Depends(get_unit_service),
    owned_by: UUID | None = Depends(get_owned_by),
) -> Page[Unit_Read]:
    # `owned` filters server-side on purpose. Narrowing a page client-side hides
    # owned units that fall on other pages and makes every derived count wrong:
    # filtering and pagination have to happen on the same side. The `owned` query
    # parameter is declared by `get_owned_by`, which also resolves the caller.
    items = service.list_units(
        faction_id=faction_id,
        subfaction_id=subfaction_id,
        q=q,
        owned_by=owned_by,
        limit=page.limit,
        offset=page.offset,
    )
    total = service.count_units(faction_id=faction_id, subfaction_id=subfaction_id, q=q, owned_by=owned_by)
    return paginate(items, total, page)


@router.get("/facets", response_model=UnitFacets)
def unit_facets(
    subfaction_id: UUID | None = None,
    q: str | None = Query(default=None, description="case-insensitive name search"),
    service: UnitService = Depends(get_unit_service),
    owned_by: UUID | None = Depends(get_owned_by),
) -> UnitFacets:
    # Declared before `/{unit_id}` so the literal path wins over the UUID param.
    # Takes `owned` for the same reason the list does: a filtered list beside an
    # unfiltered rail reads as more broken than no filter at all.
    total, by_faction = service.faction_facets(subfaction_id=subfaction_id, q=q, owned_by=owned_by)
    return UnitFacets(total=total, by_faction=by_faction)


@router.get("/{unit_id}", response_model=Unit_Read)
def get_unit(unit_id: UUID, service: UnitService = Depends(get_unit_service)) -> Unit_Read:
    return service.get_unit(unit_id)


@router.patch(
    "/{unit_id}",
    response_model=Unit_Read,
    dependencies=[Depends(get_current_admin)],
)
def update_unit(
    unit_id: UUID,
    payload: Unit_Update,
    service: UnitService = Depends(get_unit_service),
) -> Unit_Read:
    return service.update_unit(unit_id, **payload.model_dump(exclude_unset=True))


@router.delete(
    "/{unit_id}",
    status_code=status.HTTP_204_NO_CONTENT,
    dependencies=[Depends(get_current_admin)],
)
def delete_unit(unit_id: UUID, service: UnitService = Depends(get_unit_service)) -> Response:
    service.delete_unit(unit_id)
    return Response(status_code=status.HTTP_204_NO_CONTENT)


@router.post(
    "/{unit_id}/weapons",
    response_model=Unit_Read,
    dependencies=[Depends(get_current_admin)],
)
def link_weapon(
    unit_id: UUID,
    payload: WeaponLink,
    service: UnitService = Depends(get_unit_service),
) -> Unit_Read:
    return service.link_weapon(unit_id, payload.weapon_id)


@router.post(
    "/{unit_id}/abilities",
    response_model=Unit_Read,
    dependencies=[Depends(get_current_admin)],
)
def link_ability(
    unit_id: UUID,
    payload: AbilityLink,
    service: UnitService = Depends(get_unit_service),
) -> Unit_Read:
    return service.link_ability(unit_id, payload.ability_id)


@router.delete(
    "/{unit_id}/weapons/{weapon_id}",
    status_code=status.HTTP_204_NO_CONTENT,
    dependencies=[Depends(get_current_admin)],
)
def unlink_weapon(
    unit_id: UUID,
    weapon_id: UUID,
    service: UnitService = Depends(get_unit_service),
) -> Response:
    service.unlink_weapon(unit_id, weapon_id)
    return Response(status_code=status.HTTP_204_NO_CONTENT)


@router.delete(
    "/{unit_id}/abilities/{ability_id}",
    status_code=status.HTTP_204_NO_CONTENT,
    dependencies=[Depends(get_current_admin)],
)
def unlink_ability(
    unit_id: UUID,
    ability_id: UUID,
    service: UnitService = Depends(get_unit_service),
) -> Response:
    service.unlink_ability(unit_id, ability_id)
    return Response(status_code=status.HTTP_204_NO_CONTENT)