# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Warhammer Unit Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/units": {
      "post": {
        "tags": [
          "units"
        ],
        "summary": "Create Unit",
        "operationId": "create_unit_api_v1_units_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Unit_Create"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unit_Read"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "units"
        ],
        "summary": "List Units",
        "operationId": "list_units_api_v1_units_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "faction_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Faction Id"
            }
          },
          {
            "name": "subfaction_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subfaction Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "case-insensitive name search",
              "title": "Q"
            },
            "description": "case-insensitive name search"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "owned",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "only units in the caller's inventory",
              "default": false,
              "title": "Owned"
            },
            "description": "only units in the caller's inventory"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Unit_Read_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/units/facets": {
      "get": {
        "tags": [
          "units"
        ],
        "summary": "Unit Facets",
        "operationId": "unit_facets_api_v1_units_facets_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subfaction_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subfaction Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "case-insensitive name search",
              "title": "Q"
            },
            "description": "case-insensitive name search"
          },
          {
            "name": "owned",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "only units in the caller's inventory",
              "default": false,
              "title": "Owned"
            },
            "description": "only units in the caller's inventory"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitFacets"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/units/{unit_id}": {
      "get": {
        "tags": [
          "units"
        ],
        "summary": "Get Unit",
        "operationId": "get_unit_api_v1_units__unit_id__get",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Unit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unit_Read"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "units"
        ],
        "summary": "Update Unit",
        "operationId": "update_unit_api_v1_units__unit_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Unit Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Unit_Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unit_Read"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "units"
        ],
        "summary": "Delete Unit",
        "operationId": "delete_unit_api_v1_units__unit_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Unit Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/units/{unit_id}/weapons": {
      "post": {
        "tags": [
          "units"
        ],
        "summary": "Link Weapon",
        "operationId": "link_weapon_api_v1_units__unit_id__weapons_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Unit Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeaponLink"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unit_Read"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/units/{unit_id}/abilities": {
      "post": {
        "tags": [
          "units"
        ],
        "summary": "Link Ability",
        "operationId": "link_ability_api_v1_units__unit_id__abilities_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Unit Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbilityLink"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unit_Read"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/units/{unit_id}/weapons/{weapon_id}": {
      "delete": {
        "tags": [
          "units"
        ],
        "summary": "Unlink Weapon",
        "operationId": "unlink_weapon_api_v1_units__unit_id__weapons__weapon_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Unit Id"
            }
          },
          {
            "name": "weapon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Weapon Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/units/{unit_id}/abilities/{ability_id}": {
      "delete": {
        "tags": [
          "units"
        ],
        "summary": "Unlink Ability",
        "operationId": "unlink_ability_api_v1_units__unit_id__abilities__ability_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Unit Id"
            }
          },
          {
            "name": "ability_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ability Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/me/armies/{army_id}/units": {
      "post": {
        "tags": [
          "armies"
        ],
        "summary": "Add Unit",
        "operationId": "add_unit_api_v1_me_armies__army_id__units_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "army_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Army Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArmyUnitAdd"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArmyUnit_Read"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/armies/{army_id}/units/{unit_id}": {
      "patch": {
        "tags": [
          "armies"
        ],
        "summary": "Set Amount",
        "operationId": "set_amount_api_v1_me_armies__army_id__units__unit_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Unit Id"
            }
          },
          {
            "name": "army_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Army Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AmountSet"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArmyUnit_Read"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "armies"
        ],
        "summary": "Remove Unit",
        "operationId": "remove_unit_api_v1_me_armies__army_id__units__unit_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Unit Id"
            }
          },
          {
            "name": "army_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Army Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbilityLink": {
        "properties": {
          "ability_id": {
            "type": "string",
            "format": "uuid",
            "title": "Ability Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ability_id"
        ],
        "title": "AbilityLink"
      },
...
      "ArmyUnitAdd": {
        "properties": {
          "unit_id": {
            "type": "string",
            "format": "uuid",
            "title": "Unit Id"
          },
          "amount": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 1.0,
            "title": "Amount",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "unit_id"
        ],
        "title": "ArmyUnitAdd"
      },
      "ArmyUnit_Read": {
        "properties": {
          "unit": {
            "$ref": "#/components/schemas/Unit_Read"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "unit",
          "amount"
        ],
        "title": "ArmyUnit_Read"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_Unit_Read_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Unit_Read"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[Unit_Read]"
      },
...
      "UnitFacets": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "by_faction": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "format": "uuid"
            },
            "type": "object",
            "title": "By Faction"
          }
        },
        "type": "object",
        "required": [
          "total",
          "by_faction"
        ],
        "title": "UnitFacets"
      },
      "Unit_Create": {
        "properties": {
          "faction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Faction Id"
          },
          "unit_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Unit Name"
          },
          "movement": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Movement"
          },
          "toughness": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Toughness"
          },
          "armor_save": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Armor Save"
          },
          "wounds": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Wounds"
          },
          "leadership": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Leadership"
          },
          "objective_control": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Objective Control"
          },
          "points": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Points"
          },
          "invulnerable_save": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2147483647.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Invulnerable Save"
          },
          "subfaction_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subfaction Id"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "faction_id",
          "unit_name",
          "movement",
          "toughness",
          "armor_save",
          "wounds",
          "leadership",
          "objective_control",
          "points"
        ],
        "title": "Unit_Create"
      },
      "Unit_Read": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "unit_name": {
            "type": "string",
            "title": "Unit Name"
          },
          "faction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Faction Id"
          },
          "subfaction_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subfaction Id"
          },
          "movement": {
            "type": "integer",
            "title": "Movement"
          },
          "toughness": {
            "type": "integer",
            "title": "Toughness"
          },
          "armor_save": {
            "type": "integer",
            "title": "Armor Save"
          },
          "wounds": {
            "type": "integer",
            "title": "Wounds"
          },
          "invulnerable_save": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invulnerable Save"
          },
          "leadership": {
            "type": "integer",
            "title": "Leadership"
          },
          "objective_control": {
            "type": "integer",
            "title": "Objective Control"
          },
          "points": {
            "type": "integer",
            "title": "Points"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          },
          "weapons": {
            "items": {
              "$ref": "#/components/schemas/Weapon_Read"
            },
            "type": "array",
            "title": "Weapons",
            "default": []
          },
          "abilities": {
            "items": {
              "$ref": "#/components/schemas/Ability_Read"
            },
            "type": "array",
            "title": "Abilities",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "unit_name",
          "faction_id",
          "subfaction_id",
          "movement",
          "toughness",
          "armor_save",
          "wounds",
          "invulnerable_save",
          "leadership",
          "objective_control",
          "points",
          "keywords"
        ],
        "title": "Unit_Read"
      },
...
      "WeaponLink": {
        "properties": {
          "weapon_id": {
            "type": "string",
            "format": "uuid",
            "title": "Weapon Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "weapon_id"
        ],
        "title": "WeaponLink"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/login"
          }
        }
      }
    }
  }
}